+++ addOns/graaljs/CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Changed
+- Update the Active script template to contain a `scanHost` function that is called once per host being scanned.
+- Update minimum scripts add-on version to 45.15.0.
 
 ## [0.10.0] - 2025-10-07
 ### Changed